refactor(mir): carve the projected-payload provenance types out of lower
The origin classification, its provenance record and the scrutinee identity it
names are one concern and were pushing `lower/mod.rs` over its line ceiling.
